Fix memory load with limited bit width (#48)
This change fixes `load` instructions not respecting bit width parts. For example, `i32.load8_u` was loading all i32 bytes instead of a single byte.

Also, the page size was specified as 6K, while it's 64K according to the spec.

This increases the amount of spec tests passing from 25% in `master` to 33% in this branch.