Convert `Value` class hierarchy to value types (#58)
This refactors some substantial amount of code. The benefit is that it's easier to avoid ARC overhead with reduced use of reference types.

Additionally, this makes easier to implement `Equatable` on `Value` types property without utilizing hashes and avoiding hash collisions, which allows us to fix a good amount of tests. Ratio of passing tests has increased from 41% to 44%.