Use word size for register type
Use of smaller integer types requires casting it to Int for
dereferencing stack values. For example, when reading a register value
from `struct { reg, reg }`, we need to cast the register value by making
it before using it as an index to the stack because the high/low 48 bits
might be used for other purposes.