CMake: Enable WMO for Release builds
The new interpreter heavily benefits from WMO and it's even slower than
the old one without it. This change enables WMO for Release builds for
CMake builds as it's not enabled by default.

Use `CMAKE_Swift_COMPILATION_MODE` if CMake 3.29 or newer is used, and
fall back to `add_compile_options` if older.
https://cmake.org/cmake/help/latest/prop_tgt/Swift_COMPILATION_MODE.html