Translator: Cap the constant slot size to avoid slot explosion
Some large functions have a lot of constants than we expected. For
example, `SwiftSyntax.SyntaxRewriter.dispatchVisit(_:)` is over 70kb in
wasm binary size and has 1000+ unique constants, so the constant slot
size was over 1000. This caused a call to the function to consume a lot
of stack space, even though the function only accesses a few of them in
a single function call. This patch caps the constant slot size to 128
to avoid slot explosion.