Fix most of the errors when building WasmKit for WASI
Wasm is the only unsupported platform for WasmKit in the Swift Package Index build matrix.

One possible use case is when an implementation for a future Wasm feature is emulated with older Wasm features, when running WasmKit itself in an older Wasm engine.

The only remaining roadblock is the lack of `swiftasynccall`, which prevents `_CWasmKit` module from building, but the rest of the modules seem to build fine. For proper testing we also need to fix `swift test` in the Swift toolchain.