Fix breakpoints in loops & recursive functions (#237)
Currently, when continuing from a breakpoint, the debugger needs to remove it and restore the original instruction before resuming. The conventionally expected behaviour is for the original breakpoint to be preserved. For that, we don't fully resume, but only step to the next instruction, restore the original breakpoint, and then finally resume again.

This still doesn't account for branching instructions, in the same way that the current instruction stepping does not. To be resolved in a future PR, as that requires precise instruction control flow analysis.